requirements.txt
apache-airflow==2.10.4
boto3==1.35.87
Faker==33.1.0
pandas==2.2.3
pyarrow==18.1.0
